feat(alert-dialog-template): Add new component

+++ src/pages/edit-group-dialog/EditGroupDialog.jsx
@@ -1,8 +1,9 @@
-import React, { useState, useContext } from 'react';
+import React, { useState, useContext, useEffect } from 'react';
 import { Button } from '@material-ui/core';
 import useStyles from './EditGroupDialog.styles';
 import refreshDataContext from '../../stores/refreshDataStore';
 import EditableGroupDialogTemplate from '../../components/editable-group-dialog-template/EditableGroupDialogTemplate';
+import AlertDialogTemplate from '../../components/alert-dialog-template/AlertDialogTemplate';
 import GroupsService from '../../services/GroupsService';
 
 const getNestedGroupCopy = (group) => {
@@ -19,23 +20,39 @@ const EditGroupDialog = ({
 }) => {
   const classes = useStyles();
   const refreshData = useContext(refreshDataContext);
+  const [openAlertSaveDialog, setAlertSaveDialog] = useState(false);
+  const [dialogSaveAnswer, setDialogSaveAnswer] = useState(undefined);
+  const [openAlertDeleteDialog, setAlertDeleteDialog] = useState(false);
+  const [dialogDeleteAnswer, setDialogDeleteAnswer] = useState(undefined);
+
   const [newGroup, setNewGroup] = useState(getNestedGroupCopy(group));
 
-  const handleSave = async () => {
-    // TODO: Add ensuring message
-    // TODO: Add validation
+  useEffect(async () => {
     // TODO: Add loader
-    await GroupsService.updateGroup(group, newGroup);
-    refreshData();
-    onClose();
+    if (dialogSaveAnswer === 'agree') {
+      await GroupsService.updateGroup(group, newGroup);
+      refreshData();
+      onClose();
+    }
+  }, [dialogSaveAnswer]);
+
+  useEffect(async () => {
+    // TODO: Add loader
+    if (dialogDeleteAnswer === 'agree') {
+      await GroupsService.deleteGroup(group._id);
+      refreshData();
+      onClose();
+    }
+  }, [dialogDeleteAnswer]);
+
+  const handleSave = () => {
+    if (newGroup.name && newGroup.description && newGroup.users.length > 1) {
+      setAlertSaveDialog(true);
+    }
   };
 
-  const handleDeleteGroup = async () => {
-    // TODO: Add ensuring message
-    // TODO: Add loader
-    await GroupsService.deleteGroup(group._id);
-    refreshData();
-    onClose();
+  const handleDeleteGroup = () => {
+    setAlertDeleteDialog(true);
   };
 
   const dialogActions = () => (
@@ -65,13 +82,27 @@ const EditGroupDialog = ({
   );
 
   return (
-    <EditableGroupDialogTemplate
-      newGroup={newGroup}
-      setNewGroup={setNewGroup}
-      actions={dialogActions()}
-      open={open}
-      onClose={onClose}
-    />
+    <>
+      <EditableGroupDialogTemplate
+        newGroup={newGroup}
+        setNewGroup={setNewGroup}
+        actions={dialogActions()}
+        open={open}
+        onClose={onClose}
+      />
+      <AlertDialogTemplate
+        message="אתה בטוח שאתה רוצה לשמור?"
+        open={openAlertSaveDialog}
+        onClose={() => setAlertSaveDialog(false)}
+        handleAnswer={(answer) => setDialogSaveAnswer(answer)}
+      />
+      <AlertDialogTemplate
+        message="אתה בטוח שאתה רוצה למחוק?"
+        open={openAlertDeleteDialog}
+        onClose={() => setAlertDeleteDialog(false)}
+        handleAnswer={(answer) => setDialogDeleteAnswer(answer)}
+      />
+    </>
   );
 };
